old_deps/requirements.txt
Flask==0.10.1
Werkzeug==0.9.6
Jinja2==2.7.3
itsdangerous==0.24
PyYAML==3.12
requests==2.6.0
lxml==3.4.4
paramiko==1.15.2
httplib2==0.9.2
defusedxml==0.4.1
Pillow==2.3.0
